# openapi/openapi.yaml
openapi: "3.0.0"
info:
  version: 0.1.5
  title: CRM API
  description: A main CRM Api
  contact:
    name: API Team
servers:
  - url: "localhost:8080"
    description: Dev server
...
paths:
...
  /profile/photo:
    patch:
      tags:
        - profile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Upload user photo to s3
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/ProfilePhotoDTO"

    delete:
      description: Delete user photo from s3
      tags:
        - profile
      responses:
        200:
          description: Ok
        401:
          description: Unauthorized
...
components:
...
  schemas:
...
    ProfilePhotoDTO:
      x-go-type: dto.ProfilePhotoDTO
      x-go-type-import:
        name: ProfilePhotoDTOSurveyDTO
        path: github.com/krisch/crm-backend/dto
      type: object
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []